cardano-tracer: factor out `getListOfMetrics` to RTView component
