+++ cardano-tracer/src/Cardano/Tracer/Handlers/RTView/Update/Chain.hs
@@ -6,8 +6,8 @@ module Cardano.Tracer.Handlers.RTView.Update.Chain
   ( updateBlockchainHistory
   ) where
 
-import           Cardano.Tracer.Handlers.Metrics.Utils
 import           Cardano.Tracer.Handlers.RTView.State.Historical
+import           Cardano.Tracer.Handlers.RTView.Utils
 import           Cardano.Tracer.Types
 
 import           Data.Text.Read (double)
