+++ cardano-tracer/src/Cardano/Tracer/Handlers/RTView/Update/Historical.hs
@@ -15,7 +15,6 @@ module Cardano.Tracer.Handlers.RTView.Update.Historical
   ) where
 
 import           Cardano.Tracer.Environment
-import           Cardano.Tracer.Handlers.Metrics.Utils
 import           Cardano.Tracer.Handlers.RTView.State.Historical
 import           Cardano.Tracer.Handlers.RTView.State.Last
 import           Cardano.Tracer.Handlers.RTView.Update.Chain
