+++ cardano-tracer/src/Cardano/Tracer/Handlers/RTView/Update/Resources.hs
@@ -7,9 +7,9 @@ module Cardano.Tracer.Handlers.RTView.Update.Resources
   ( updateResourcesHistory
   ) where
 
-import           Cardano.Tracer.Handlers.Metrics.Utils
 import           Cardano.Tracer.Handlers.RTView.State.Historical
 import           Cardano.Tracer.Handlers.RTView.State.Last
+import           Cardano.Tracer.Handlers.RTView.Utils
 import           Cardano.Tracer.Handlers.Utils
 import           Cardano.Tracer.Types
