+++ cardano-tracer/src/Cardano/Tracer/Handlers/RTView/Update/Transactions.hs
@@ -6,8 +6,8 @@ module Cardano.Tracer.Handlers.RTView.Update.Transactions
   ( updateTransactionsHistory
   ) where
 
-import           Cardano.Tracer.Handlers.Metrics.Utils
 import           Cardano.Tracer.Handlers.RTView.State.Historical
+import           Cardano.Tracer.Handlers.RTView.Utils
 import           Cardano.Tracer.Types
 
 import           Data.Text.Read (decimal)
